+++ CHANGELOG.md
@@ -40,6 +40,7 @@
     - Added context menu option "Scene lock" to prevent accidental changes
 - Module [TRANSIT](./docs/Transit.md)
     - Added context menu option for unbinding all bound parameters of a module (#268)
+    - Added "Auto"-mode besides "Read" and "Write" ([manual](./docs/Transit.md#auto-mode)) (#269)
 
 ## 1.9.0